Address restore review edge cases
